requirements.txt
boto3==1.26.139
langchain==0.0.194
langchain_serve==0.0.37
numpy==1.23.5
openai==0.27.6
pandas==1.5.3
pypdf==3.9.1
python-dotenv==1.0.0
Requests==2.31.0
spacy==3.5.3
streamlit==1.22.0
